# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "创意喵平台 API",
    "description": "创意喵是一个连接商家和创作者的任务撮合平台 API",
    "version": "1.0.0",
    "contact": {
      "name": "创意喵"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "开发环境"
    }
  ],
...
  "paths": {
    "/auth/wechat-mini-login": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "微信小程序登录",
        "description": "通过微信小程序wx.login返回的code进行登录，自动创建新用户或返回已存在用户",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code"
                ],
                "properties": {
                  "code": {
                    "type": "string",
                    "description": "wx.login返回的code"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WechatMiniLoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "tags": [
          "任务"
        ],
        "summary": "获取可认领任务列表",
        "description": "公开接口，无需认证",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "created_at",
                "price_desc",
                "price_asc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "tags": [
          "任务"
        ],
        "summary": "获取任务详情",
        "description": "公开接口，无需认证",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": [
          "用户"
        ],
        "summary": "获取当前用户信息",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "用户"
        ],
        "summary": "更新用户资料",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/users/credits": {
      "get": {
        "tags": [
          "用户"
        ],
        "summary": "获取用户信用分详情",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/mine/stats": {
      "get": {
        "tags": [
          "我的"
        ],
        "summary": "获取我的页统计",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/user/profile": {
      "get": {
        "tags": [
          "用户"
        ],
        "summary": "获取用户个人资料",
        "description": "获取当前登录用户的详细个人资料信息",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "用户"
        ],
        "summary": "更新用户个人资料",
        "description": "更新当前登录用户的个人资料信息",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string",
                    "description": "昵称",
                    "maxLength": 50
                  },
                  "phone": {
                    "type": "string",
                    "description": "手机号",
                    "pattern": "^1[3-9]\\d{9}$"
                  },
                  "avatar": {
                    "type": "string",
                    "description": "头像URL"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user/password": {
      "put": {
        "tags": [
          "用户"
        ],
        "summary": "修改密码",
        "description": "修改当前登录用户的密码",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "old_password",
                  "new_password"
                ],
                "properties": {
                  "old_password": {
                    "type": "string",
                    "description": "原密码",
                    "minLength": 6
                  },
                  "new_password": {
                    "type": "string",
                    "description": "新密码",
                    "minLength": 6,
                    "maxLength": 50
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "修改成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误或原密码错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/business/tasks/{id}": {
      "delete": {
        "tags": [
          "任务"
        ],
        "summary": "取消任务",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/business/claim/{id}": {
      "get": {
        "tags": [
          "认领"
        ],
        "summary": "获取认领详情",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/works": {
      "get": {
        "tags": [
          "任务"
        ],
        "summary": "获取过审作品列表（灵感画廊）",
        "description": "公开接口，无需认证。返回状态为 approved 的认领作品，每条记录包含 materials 媒体文件列表。",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "created_at"
              ],
              "default": "created_at"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Work"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/works/{id}": {
      "get": {
        "tags": [
          "任务"
        ],
        "summary": "获取作品详情",
        "description": "公开接口，无需认证。返回单条过审认领作品，包含 materials 媒体文件列表。",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Work"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "作品不存在或未通过审核"
          }
        }
      }
    },
    "/creator/tasks": {
      "get": {
        "tags": [
          "任务"
        ],
        "summary": "创作者获取任务列表",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/creator/claim": {
      "post": {
        "tags": [
          "认领"
        ],
        "summary": "认领任务",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "task_id"
                ],
                "properties": {
                  "task_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "认领成功"
          }
        }
      }
    },
    "/creator/claims": {
      "get": {
        "tags": [
          "认领"
        ],
        "summary": "获取我的认领列表",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/creator/claim/{id}/submit": {
      "put": {
        "tags": [
          "认领"
        ],
        "summary": "提交交付",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "作品链接或内容描述"
                  },
                  "materials": {
                    "type": "array",
                    "description": "媒体文件列表（先通过 /upload 接口上传，获得路径后传入）",
                    "items": {
                      "type": "object",
                      "required": [
                        "file_name",
                        "file_path",
                        "file_type"
                      ],
                      "properties": {
                        "file_name": {
                          "type": "string",
                          "description": "文件名"
                        },
                        "file_path": {
                          "type": "string",
                          "description": "文件路径（上传后返回的路径）"
                        },
                        "file_size": {
                          "type": "integer",
                          "description": "文件大小（字节）"
                        },
                        "file_type": {
                          "type": "string",
                          "description": "文件类型，如 video/mp4, image/jpeg"
                        },
                        "thumbnail_path": {
                          "type": "string",
                          "description": "缩略图路径（可选）"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "提交成功"
          }
        }
      }
    },
    "/creator/wallet": {
      "get": {
        "tags": [
          "钱包"
        ],
        "summary": "获取钱包信息",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/creator/transactions": {
      "get": {
        "tags": [
          "钱包"
        ],
        "summary": "获取交易记录",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/creator/stats": {
      "get": {
        "tags": [
          "钱包"
        ],
        "summary": "获取创作者统计",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/creator/chart/income": {
      "get": {
        "tags": [
          "钱包"
        ],
        "summary": "获取收益图表",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "7d",
                "30d",
                "90d"
              ],
              "default": "7d"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
    "/business/tasks/{id}/claims": {
      "get": {
        "tags": [
          "认领"
        ],
        "summary": "获取任务认领列表",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/business/claims": {
      "get": {
        "tags": [
          "认领"
        ],
        "summary": "获取所有认领列表",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/business/claim/{id}/review": {
      "put": {
        "tags": [
          "认领"
        ],
        "summary": "验收认领",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "result"
                ],
                "properties": {
                  "result": {
                    "type": "integer",
                    "description": "1-通过，0-退回"
                  },
                  "comment": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "验收成功"
          }
        }
      }
    },
...
    "/business/recharge": {
      "post": {
        "tags": [
          "钱包"
        ],
        "summary": "充值",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount"
                ],
                "properties": {
                  "amount": {
                    "type": "number"
                  },
                  "payment_method": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "充值成功"
          }
        }
      }
    },
    "/business/stats": {
      "get": {
        "tags": [
          "钱包"
        ],
        "summary": "获取商家统计",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token"
      },
      "AdminAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token",
        "description": "管理员 Token"
      }
    },
    "schemas": {
      "WechatMiniLoginResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "成功"
          },
          "data": {
            "type": "object",
            "properties": {
              "token": {
                "type": "string"
              },
              "user": {
                "$ref": "#/components/schemas/User"
              },
              "is_new": {
                "type": "boolean",
                "description": "是否为新创建的用户"
              }
            }
          }
        }
      },
...
      "TaskDetailResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Task"
          }
        }
      },
...
    }
  }
}